% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-broadband.R
\docType{data}
\name{broadband}
\alias{broadband}
\title{Broadband Adoption & Peak Speed by Country for Q3 2013}
\format{A data frame with 240 rows and 8 variables:
\describe{
  \item{CountryCode}{FILL}
  \item{Year}{FILL}
  \item{Quarter}{FILL}
  \item{PeakMbps}{FILL}
  \item{Adoption_BB}{FILL}
  \item{Adoption_LowBB}{FILL}
  \item{Adoption_HighBB}{FILL}
  \item{Count_UniqueIP}{FILL}
}}
\source{
Leading Provider of Internet Statistics
}
\usage{
broadband
}
\description{
A dataset that contains peak speeds and percentage adoption of Broadband, 
Narrow Broadband (slower), and High speed Broadband rates by Country Code 
for Q3 2013 only. Note this data was scraped from the website of a leading 
provider of internet statistics and trends.
}
\keyword{datasets}
