% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-broadbandQtr.R
\docType{data}
\name{broadbandQtr}
\alias{broadbandQtr}
\title{Broadband Adoption & Peak Speed by Country and Quarter}
\format{A data frame with 3,775 rows and 7 variables:
\describe{
  \item{COUNTRY}{FILL}
  \item{Period}{FILL}
  \item{Adoption_HighBB}{FILL}
  \item{Adoption_LowBB}{FILL}
  \item{Adoption_BB}{FILL}
  \item{AveSpeed}{FILL}
  \item{AvePeakSpeed}{FILL}
}}
\source{
Leading Provider of Internet Statistics
}
\usage{
broadbandQtr
}
\description{
A dataset that contains Broadband adoption and speed through multiple quarters
by country. Note this data was scraped from the website of a leading 
provider of internet statistics and trends.
}
\keyword{datasets}

