% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDT.R
\name{checkdt}
\alias{checkdt}
\title{Internal Helper Functions}
\usage{
checkdt(dt = NULL, cols = NULL, noerror = FALSE)
}
\arguments{
\item{dt}{An object to make common checks to verify it is a valid data.table}

\item{cols}{An optional character vector of column names to check against column names of dt}

\item{noerror}{A boolean where if TRUE, checkdt will return a numeric error code 
instead of a generated error that results in code stoppage. Default value is FALSE}
}
\value{
Returns an error code if flag \code{noerror} is TRUE. Otherwise returns NULL or 
     code execution is stopped if one of the checks fail. See details for the checks 
     that are performed.
}
\description{
\code{checkdt} is intended for developers who write functions that 
     accept a \code{data.table} as an input. \code{checkdt} performs common checks 
     and generates accessible and informative errors (or returns error codes if 
     desired), if a data.table fails various checks. An optional character vector 
     argument \code{cols} enables an additional check to determine if they exist as 
     column names of \code{dt}. See details below the checks that are performed, the
     error messages, and if applicable, what error codes are returned.
 @details If any of the following are true, a friendly error will be shown and the 
     code execution will be stopped:
         \itemize{
             \item{dt is NULL}
             \item{dt is not of class data.table (data.frame)}
             \item{dt has 0 rows}
             \item{if \code{cols} not valid column names of dt (optional check)}
         }
    The following error codes are generated when each of the above situations 
    occur, \strong{and the flag \code{noerror} is set to TRUE:}
         \itemize{
             \item{if null: 999} 
             \item{if class not data.table: 101} 
             \item{if empty: 555} 
             \item{if cols not column names: 422} 
         }
}
\examples{
\dontrun{
#### dt is NULL
checkdt(dt = NULL) # Prints "Error: no data.table provided"
checkdt(dt = NULL, noerror = TRUE) # Returns 999

#### dt not class data.table(data.table)
checkdt(dt = "test") # Prints "Error: dt not of class data.table or data.frame"
checkdt(dt = "test", noerror = TRUE) # Returns 101

#### dt has 0 rows
checkdt(dt = data.table()) # Prints "Error: no data in dt"
checkdt(dt = data.table(), noerror = TRUE) # Returns 555

#### check columns also
DT <- data.table(X = 1, Y = 2)

checkdt(dt = DT, cols = c("A", "B")) # Prints "Error: the following cols not in data.table: A, B"
checkdt(dt = DT, noerror = TRUE) # Returns 422
}
}

