% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{CleanCols}
\alias{CleanCols}
\alias{xDate}
\title{Miscellaneous Functions}
\usage{
CleanCols(i)

xDate(DateVector)
}
\arguments{
\item{i}{A data table to clean}

\item{DateVector}{A vector of values of class POSIXct}
}
\description{
Miscellaneous Functions
}
\section{Functions}{
\itemize{
\item \code{CleanCols}: A function to remove all columns that have only NA values

\item \code{xDate}: A function to expand elements of a timestamp into 
numerous columns in a new data.table
}}

