% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{dataclean}
\alias{CleanCols}
\alias{RoundCols}
\alias{dataclean}
\alias{ddup}
\alias{p_setcolorder}
\alias{xDate}
\title{Functions to help with Data Cleaning}
\usage{
CleanCols(i)

RoundCols(i, digits = 2, igncols = NULL, cols = NULL, b_skip = FALSE,
  b_copy = TRUE)

p_setcolorder(i, cols = NULL, aslast = TRUE)

xDate(date_vec)

ddup(i, cols = NULL, verbose = TRUE)
}
\arguments{
\item{i}{A data table to operate on}

\item{digits}{A numeric indicating the number of digits to round in \code{RoundCols}}

\item{igncols}{A character vector indicating which columns to ignore}

\item{cols}{columns of a data.table to remove duplicate rows accross}

\item{b_skip}{A boolean indicating whether to skip non-numeric columns that \code{RoundCols} encounters}

\item{b_copy}{A boolean indicating whether to make a copy of the data, or operate on by reference}

\item{aslast}{A boolean indicating whether arrange the non-order columns before or after (asLast = TRUE)
the ordered ones}

\item{date_vec}{A vector of values of class POSIXct}

\item{verbose}{A boolean indicating whether to print information on the console}
}
\description{
Functions to help with Data Cleaning
}
\section{Functions}{
\itemize{
\item \code{CleanCols}: A function to remove all columns that have only NA values

\item \code{RoundCols}: A convienience wrapper for \code{round} that applies to all or a subset of cols

\item \code{p_setcolorder}: A convienience wrapper for \code{data.table::setcolorder} that makes it easy 
set the order of a subset of columns

\item \code{xDate}: A function to expand elements of a timestamp into 
numerous columns in a new data.table

\item \code{ddup}: A function to remove duplicates across all columns (default),
or a given set of columns
}}

