% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateMorph.R
\name{DateMorph}
\alias{DateMorph}
\title{Functions to Change Column Classes}
\usage{
DateMorph(dt, cols = NULL, format = NULL, tz = c("current", "utc", "gmt"),
  copy = FALSE)
}
\arguments{
\item{dt}{A data.table to morph column classes}

\item{cols}{An optional character vector of column names to operate on.
Not applicable for function ClassMorph.}

\item{format}{Format of the value to convert to date. See examples for acceptable formats.}

\item{tz}{A character value indicating the timezone. Options are: "current", "utc", and "gmt"}

\item{copy}{A boolean value indicating whether to alter \code{dt} in memory or 
whether to return a \emph{new (copy)} of the input data.table.}
}
\description{
The functions \code{\link{ClassMorph}}, and \code{\link{NumMorph}}, in addition to \code{DateMorph},
are functions designed to make dealing with column classes easy.
}
\section{Important}{

This function is deprecated while development of a complete rehaul is ongoing.
The new \code{DateMorph} has an algorithm to smartly identify date columns anywhere 
in the table, while also identifying the format of the dates, using \code{fudate} 
to do both, only then changing the column class to the appropriate date class
}
\examples{
library(data.table)

#-------------------------------------------------------------------------------------
# Example: Change column classes in batch
#-------------------------------------------------------------------------------------

# Create example data
dt <- data.table(Int     = as.factor(c(1:2, "-", 4:5)),
                 Double  = rep(2.0, 5),
                 Factor  = as.factor(c(1, "garbage", 2, "", 3)),
                 nFactor = as.factor(rep(999, 5)),
                 Char    = letters[1:5],
                 Date    = rep("2020-01-22", 5))
dt
#-------------------------------------------------------------------------------------
# Errors when generating NA
#
# The following line will generate NAs because there is an element called 
# "garbage" in the column Factor. In fact, a missed non-numeric value in
# an otherwise mostly number column is the main reason the class automatically
# gets read as a factor. 
pclass(dt)
ClassMorph(dt, "factor", "integer")

#-------------------------------------------------------------------------------------
# The 'force' flag
#
# Naturally, when converting "garbage" to an integer, you will get an NA.
# To prevent accidental loss of data, the default behavior is to generate
# an error when conversion creates NAs. This should only be expected to 
# happen when going from factor to numeric/integer. In cases where NAs are expected,
# set the force flag = TRUE:
#
ClassMorph(dt, "factor", "integer", force = TRUE) # no error
pclass(dt)

#-------------------------------------------------------------------------------------
# The 'copy' flag
#
# Not modifying dt in place, rather generate a new copy of dt.
# Notice in the previous examples, ClassMorph had no return object. That is due 
# to the fact that ClassMorph was modifying the data in-place. Traditionally
# r has been a "copy-on-modify" language. That is typically not the case today,
# but in some cases it is desired. For example, if we want to be conservative 
# about accidental data loss, we would specify the flag 'copy=TRUE' and then
# manually delete the old table. 
#
newdt <- ClassMorph(dt, "numeric", "factor", copy = TRUE)

identical(cc(dt), cc(newdt))    # classes are not equal
pclass(dt)                      # confirm correct conversion
pclass(newdt)

#-------------------------------------------------------------------------------------
# IMPORTANT: Note that without specifying 'copy=TRUE', newdt is simply just 
# pointing to dt, as observed here:
#
pclass(dt)
newdt <- ClassMorph(dt, "numeric", "factor")

# dt and newdt are the same table in memory
identical(cc(dt), cc(newdt))
identical(dt, newdt)    




}

