% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescribeData.R
\name{describe}
\alias{describe}
\title{Describing Data}
\usage{
describe(dt)
}
\arguments{
\item{dt}{A data.table to describe}
}
\description{
These functions help "describe" data.tables in order to
     get an initial view of what it contains, the size, the column 
     classes, etc.
}
\section{Functions}{
\itemize{
\item \code{describe}: A wrapper around function \code{str}, with preset parameters
convient for objects of class data.table (data.frame)
}}
\examples{
library(data.table)

dt <- data.table(ColumnA = c(1,2,3),
                 ColumnB = c(4,5,6))
describe(dt)

# Classes 'data.table' and 'data.frame': 
#      3 obs. of 2 variables:
#  $ ColumnA: num  1 2 3
#  $ ColumnB: num  4 5 6

}

