% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring.R
\name{dtDescribe}
\alias{dtDescribe}
\title{Functions to help with pre-data exploration}
\usage{
dtDescribe(DT, all = FALSE, cols = NULL)
}
\arguments{
\item{DT}{A data.table}

\item{all}{Boolean indicating whether to perform for all columns (default is FALSE). If 
this is false and optional argument "cols" was also not provided, the default behavior 
is to perform this for only factor and character column classes in DT}

\item{cols}{optional arg providing column names to operate on}
}
\value{
A descriptive data.table
}
\description{
Helper functions to assist with basic data exploration tasks.
}
\examples{
# dtDescribe(DT)
}

