% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_describe.R
\name{easy_describe}
\alias{easy_describe}
\alias{split_by_class}
\alias{wcolNA}
\alias{wrowNA}
\title{Functions to help with pre-data exploration}
\usage{
easy_describe(DT, cols = NULL, cclass = NULL, FUN = NULL, all = NULL)

split_by_class(DT)

wcolNA(DT)

wrowNA(DT)
}
\arguments{
\item{DT}{A data.table}

\item{cols}{optional arg providing column names to operate on}

\item{cclass}{Not implemented. Useful if desiring to describe all columns a particular class}

\item{FUN}{NI; provide function to apply over cols (if provided)}

\item{all}{Boolean indicating whether to perform for all columns (default is FALSE). If 
this is false and optional argument "cols" was also not provided, the default behavior 
is to perform this for only factor and character column classes in DT}
}
\value{
A descriptive data.table
}
\description{
Helper functions to assist with basic data exploration tasks.
}
\section{Functions}{
\itemize{
\item \code{split_by_class}: A function to split the column names of a data.table into groups specified by class

\item \code{wcolNA}: Which columns contain all NA values down the rows (all obervations are NA)

\item \code{wrowNA}: Which rows contain all NA values across the columns
}}

\examples{
library(data.table)
easy_describe(as.data.table(iris))
}
