% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FiberMSADem.R
\docType{data}
\name{fiberMSADem}
\alias{fiberMSADem}
\title{GDP and Fiber Internet Access Data, by MSA}
\format{A data frame with 1,418 rows and 21 variables:
\describe{
  \item{Fips}{FILL}
  \item{Year}{FILL}
  \item{Area}{FILL}
  \item{GDP}{FILL}
  \item{GDPGrowth}{FILL}
  \item{MetroAreaSize}{FILL}
  \item{PersonalIncome_PerCap}{FILL}
  \item{JobsCount}{FILL}
  \item{JobsGrowth}{FILL}
  \item{TYPE}{FILL}
  \item{LandArea}{FILL}
  \item{Population}{FILL}
  \item{Households}{FILL}
  \item{Income_Median}{FILL}
  \item{Income_Less25}{FILL}
  \item{Income_25to50}{FILL}
  \item{Income_50to100}{FILL}
  \item{Income_100to200}{FILL}
  \item{Income_greater200}{FILL}
  \item{OpticalFiber}{FILL}
  \item{advdl_gr1gig}{FILL}
}}
\source{
BEA and NTIA
}
\usage{
fiberMSADem
}
\description{
A dataset that contains GDP and demographic data (at the MSA level) from the BEA 
merged with Fiber internet adoption and access data from NTIA
}
\keyword{datasets}

