% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fudates.R
\name{fudate}
\alias{extract_date}
\alias{fudate}
\title{Identify Dates with Patterns}
\usage{
fudate(form = c("ymd", "ydm", "mdy", "dmy"), yr4 = TRUE)

extract_date(x, form, yr4, ...)
}
\arguments{
\item{form}{A character value of 'ymd', 'ydm', 'mdy', 'dmy'. See details below for addition info}

\item{yr4}{A boolean indicating whether to expect a 4 digit or 2 digit date}

\item{x}{A vector, column, or text which contains dates to extract}

\item{...}{NOT IMPLEMENTED}
}
\description{
The function \code{fudate} returns the regex pattern that identifies any date using most common date formats.
}
\details{
A universal regex pattern to 
identify any date with certainty is very difficult to do. This regex uses built in logic to rule out 
false positives despite having the correct format. To ensure accuracy, this pattern does not attempt
to identify the position of the year, month, or day groups. 

The argument \code{form} is a character value of 'ymd', 'ydm', 'mdy', 'dmy' which determines the returned
regex pattern:

 \itemize{
     \item{\code{ymd} identifies dates where the year is first, month second, followed by day.}
     \item{The other options are: \code{ydm}, \code{mdy}, or \code{dmy}}
     \item{\code{yr4} is a boolean specifying whether to expect a 4 digit or 2 digit year. 
     This is done to ensure the highest possible accuracy in the match}
 }
}
\section{Functions}{
\itemize{
\item \code{extract_date}: A function that utilizes \code{fudate} to extract elements that are 
identified as dates
}}
\section{Additional Information}{

For details of the logic see the vignette LINK HERE. This vignette also serves as an basic 
intro to regular expressions using dates for an example case study
}

