% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveCol.R
\name{RemoveCols}
\alias{RemoveCols}
\title{Easy & flexible column deletion}
\usage{
RemoveCols(dt, byname = NULL, byval = NULL, exact = TRUE, all = TRUE,
  copy = FALSE)
}
\arguments{
\item{dt}{A data.table that contains columns to remove}

\item{byname}{An optional character vector of length 1 or greater used when wanting to remove columns 
by approximate (or exact) matching of names. If argument "exact" is TRUE (default), "byname"
can be a vector of column names to remove in batch. If false, byname should be a pattern to
match against the column names}

\item{byval}{An optional value to search through each column and find. If "all" is TRUE (default), every 
value in the column has to match "byval", else only 1 match is required to flag column for
removal. If "exact" is FALSE (default is TRUE), approximate text matching will occur on the values.}

\item{exact}{A boolean value with a default of TRUE indicating whether matching of column names or values
should be an exact match or an approximate match based on pattern matching}

\item{all}{A boolean value with a default of TRUE indicating whether ALL values in a column should match
the value provided by "byval". If FALSE, only one match is required to flag column for removal}

\item{copy}{A boolean with a default of TRUE indicating whether a copy of the data.table should be returned.
If false, the data.table is altered in memory and the modified input "dt" is returned}
}
\value{
A data.table such that some columns have been removed. If no columns were flagged for removal, 
     an error code will be returned and a friendly message will be printed on the screen
}
\description{
The function \code{RemoveCols} helps to remove columns of a data.table, data.frame, list, 
or generally any object with named elements efficiently with an intuitive interface
}

