% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easyfilter.R
\name{rFilter}
\alias{cFilter}
\alias{rFilter}
\title{Quick filtering rows or columns}
\usage{
rFilter(dt, val = NA, boolfun = NULL, cols = NULL)

cFilter(dt, val = NA, boolfun = NULL, cols = NULL)
}
\arguments{
\item{dt}{A data.table to filter rows or columns}

\item{val}{A value to filter rows or columns on}

\item{boolfun}{A function that can be used instead of val to filter. This function
must be a "predicate" function applied across rows (\code{cFilter}), or down
columns (\code{rFilter}). See details and examples below.
(returns T/F based on a test of values in dt
applied across rows, or down columns)}

\item{cols}{An OPTIONAL character vector specifying columns to include as part of
testing the conditions. If cols is not provided, all column of dt will be
tested}
}
\description{
\code{rFilter} is a function that enables quick filtering a data.table based on
 a TRUE/FALSE condition applied down the column (e.g. returns entire row if
 condition is TRUE for each value down rows of specified columns).
 \code{cFilter} is analogous for columns. This will test a condition across rows
 of every specified columns and drop each if the condition is met.
}
\section{Functions}{
\itemize{
\item \code{rFilter}: A function to filter rows of a data.table

\item \code{cFilter}: A function to filter columns of a data.table
}}

