% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterRows.R
\name{rFilter_All}
\alias{FilterRows}
\alias{rFilter}
\alias{rFilter_All}
\title{Quick filtering rows or columns}
\usage{
rFilter_All(dt, val = NA, boolfun = NULL, cols = NULL, copy = TRUE)

rFilter(dt, val = NA, exact = FALSE, set = NULL)

FilterRows(dt, val = NA, exact = FALSE, set = NULL)
}
\arguments{
\item{dt}{A data.table to filter rows or columns}

\item{val}{A value to filter rows on}

\item{boolfun}{A function that can be used instead of val to filter. This function
must be a "predicate" function applied down columns (\code{rFilter_All/rFilter}). 
See details and examples below. (A predicate function returns T/F based on a test of 
values in dt applied across rows, or down columns)}

\item{cols}{An optional character vector specifying columns to include as part of
testing the conditions. If cols is not provided, all column of dt will be 
tested}

\item{copy}{A boolean indicating whether the returned data.table should be a 
copy of "dt" or a reference to the data in memory. Default value is TRUE to protect
against unnexpected changes to the original data, however setting copy=FALSE when
appropriate is much more efficient. See example use cases.}

\item{exact}{A boolean that describes whether to exactly match. Default is FALSE which means a text based
search will occur in order to match whether val is contained in the row}

\item{set}{An optional list element that can be provided to set a desired new column after filtering.
set should be a named list where the name corresponds to the new column name, and the value
is the new column value. This is ideal when needing to quick boolean or indicator column to 
identify the group that was just filtered. See examples}
}
\description{
The functions \code{rFilter_All} and \code{rFilter} enable quick filtering a 
data.table based on a TRUE/FALSE condition applied down the column (e.g. returns 
entire row if condition is TRUE for each value down rows of specified columns).
}
\section{Functions}{
\itemize{
\item \code{rFilter_All}: A function to filter rows in a data.table based on a search
condition applied through all columns. Rows such that a match is made in every column
will be returned. For a less strict rule, see \code{rFilter}

\item \code{rFilter}: A functional similar to \code{rFilter_All} except this 
function returns rows if one column meets search criteria

\item \code{FilterRows}: NOT IMPLEMENTED
}}

