% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveread.R
\name{saver}
\alias{readr}
\alias{saver}
\title{Import/Export Compressed Data Quickly}
\usage{
saver(...)

readr(..., env = NULL)
}
\arguments{
\item{...}{Objects to save as compressed rds file or read from rds files 
into the specified environment (default is parent environment of caller)}

\item{env}{An environment to read object into (default is parent of caller)}
}
\description{
Utility functions and wrappers for saveRDS/readRDS (for internal use)
}
\section{Functions}{
\itemize{
\item \code{saver}: A function to save an r object as compressed file

\item \code{readr}: A function to read rds files and save object in specified environment
}}

