% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_index.R
\name{split_by_index}
\alias{split_by_index}
\title{Split into List}
\usage{
split_by_index(dat = NULL, index = NULL, include_first = TRUE,
  include_at_index = TRUE)
}
\arguments{
\item{dat}{A matrix, data.frame, data.table, or vector. If table form, split will occur at rows}

\item{index}{A numeric vector that identifies the positions for which to split at}

\item{include_first}{A logical indicating whether it is required to keep elements leading up to the first
index. For example, if I want to split a vector at position 5, 10, and 15, the default value of TRUE would ensure
elements 1 to 5 are returned (or 1 to 4 if argument \code{include_at_index} were set to FALSE)}

\item{include_at_index}{A logical indicating whether to include the element at the index, or drop it.
The default value of TRUE will keep the element in result}
}
\description{
Split into list element chunks using a given index that identifies the split points
}

