% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIndex.R
\name{xIndex}
\alias{xIndex}
\title{Generate Indices}
\usage{
xIndex(len = NULL, eachN, mat = FALSE)
}
\arguments{
\item{len}{A numeric value representing the total observations to split an index over}

\item{eachN}{A numeric value representing the desired number of splits}

\item{mat}{A boolean indicating whether to return the splits as a list or matrix (simplified)}
}
\description{
A function to create a split index. Comes in handy if parallelizing a computation over a very large dataset
}
\examples{
len <- 1000
eachN <- 10
xIndex(len, eachN) # Not simplified
xIndex(len, eachN, mat = TRUE) # simplified
}

