% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classmorph.R
\name{ClassMorph}
\alias{ClassMorph}
\alias{NumMorph}
\alias{cc}
\alias{pclass}
\title{Functions to Change Column Classes}
\usage{
ClassMorph(DT, old = c("factor", "integer", "character", "numeric", "logical",
  "Date"), new = c("factor", "integer", "character", "numeric"),
  copy = FALSE, force = FALSE)

NumMorph(dt, cols = NULL, copy = FALSE, verbose = FALSE)

pclass(dt)

cc(dt)
}
\arguments{
\item{DT}{A data.table to use for any "class" operations}

\item{old}{ClassMorph only. A column type, specified as a character value, to
detect and convert to the class specified by 'new'}

\item{new}{ClassMorph only. A column type, specified as a character value, to
convert all columns of the class specified by 'old' to.}

\item{copy}{A boolean value indicating whether to alter \code{dt} in memory
or whether to return a new (copy) of the input data.table.}

\item{force}{A boolean indicating whether to force conversion from class
factor to class numeric despite NAs being generated.}

\item{cols}{An optional character vector of column names to operate on. Not
applicable for function ClassMorph.}

\item{verbose}{A boolean indicating whether to be chatty}
}
\value{
Returns a data.table that is either a copy of the input dt that has
  been modified, or the same input dt that has been modified in memory
}
\description{
The functions \code{ClassMorph}, and \code{NumMorph}, in addition to
\code{DateMorph}, are functions designed to make dealing with column classes
easy. See examples for how to avoid uncessary frustrations related to
unexpected column classes or date formats.
}
\details{
This function takes care of multiple conversion steps that are sometimes
needed to avoid data loss or unexpected results (e.g. a factor that "looks"
like the number "1" may not convert to a \code{numeric} "1". To get around
this common frustration, this function will perform an indirect conversion to
\code{character} before converting to \code{numeric})
}
\section{Functions}{
\itemize{
\item \code{ClassMorph}: A function to convert all columns of class "old" to class "new". This
function handles the indirect conversions sometimes needed to avoid data loss or
unexpected results (see details).

\item \code{NumMorph}: A function to standards classes to "numeric" for all columns of
dt, where a conversion to numeric would not generate NA values.

\item \code{pclass}: A wrapper to sapply that prints the class of each column on the R console

\item \code{cc}: Similar to \code{pclass} except this function returns an object, 
instead of simply printing to the console
}}
\examples{
#-------------------------------------------------------------------------------------
# Example: Change column classes in batch
#-------------------------------------------------------------------------------------

library(data.table)
DT <- data.table(A = as.factor(c(1:2, "-", 4:5)),
                 B = rep(2.0, 5),
                 C = as.factor(c(1, "garbage", 2, "", 3)),
                 D = as.factor(rep(999, 5)),
                 E = letters[1:5],
                 F = rep("2020-01-22", 5))


# > DT
#    A B       C   D E          F
# 1: 1 2       1 999 a 2020-01-22
# 2: 2 2 garbage 999 b 2020-01-22
# 3: - 2       2 999 c 2020-01-22
# 4: 4 2         999 d 2020-01-22
# 5: 5 2       3 999 e 2020-01-22


ClassMorph(DT, "factor", "integer", force = TRUE) # no error

# Note that this converted by reference. To make copy of data in mem, 
# set copy = TRUE

DT
#     A B  C   D E          F
# 1:  1 2  1 999 a 2020-01-22
# 2:  2 2 NA 999 b 2020-01-22
# 3: NA 2  2 999 c 2020-01-22
# 4:  4 2 NA 999 d 2020-01-22
# 5:  5 2  3 999 e 2020-01-22
}

