% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{CleanCols}
\alias{CleanCols}
\alias{ddup}
\alias{xDate}
\title{Miscellaneous Functions}
\usage{
CleanCols(i)

xDate(DateVector)

ddup(i, cols = NULL, verbose = TRUE)
}
\arguments{
\item{i}{A data table to clean or dedup}

\item{DateVector}{A vector of values of class POSIXct}

\item{cols}{columns of a data.table to remove duplicate rows accross}

\item{verbose}{A boolean indicating whether to print information on the console}
}
\description{
Miscellaneous Functions
}
\section{Functions}{
\itemize{
\item \code{CleanCols}: A function to remove all columns that have only NA values

\item \code{xDate}: A function to expand elements of a timestamp into 
numerous columns in a new data.table

\item \code{ddup}: A function to remove duplicates across all columns (default),
or a given set of columns
}}

