% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DateMorph.R
\name{DateMorph}
\alias{DateMorph}
\title{Functions to Change Column Classes}
\usage{
DateMorph(dt, cols = NULL, format = NULL, tz = c("current", "utc", "gmt"),
  copy = FALSE)
}
\arguments{
\item{dt}{A data.table to morph column classes}

\item{cols}{An optional character vector of column names to operate on. Not
applicable for function ClassMorph.}

\item{format}{Format of the value to convert to date. See examples for acceptable formats.}

\item{tz}{A character value indicating the timezone. Options are: "current", "utc", and "gmt"}

\item{copy}{A boolean value indicating whether to alter \code{dt} in memory
or whether to return a new (copy) of the input data.table.}
}
\description{
The functions \code{\link{ClassMorph}}, and \code{\link{NumMorph}}, in addition to \code{DateMorph},
are functions designed to make dealing with column classes easy.
}
\section{Important}{

This function is deprecated while development of a complete rehaul is ongoing.
The new \code{DateMorph} has an algorithm to smartly identify date columns anywhere 
in the table, while also identifying the format of the dates, using \code{fudate} 
to do both, only then changing the column class to the appropriate date class
}

