% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idoutliers.R
\name{id_outliers}
\alias{id_outliers}
\title{A function that identifies outliers in a vector of numeric data}
\usage{
id_outliers(x = NULL)
}
\arguments{
\item{x}{A vector that is of class numeric, or coercible to one}
}
\value{
A logical vector where TRUE identifies the position of an outlier in the
    input argument
}
\description{
A function that identifies outliers in a vector of numeric data
}
\examples{
#vec <- rnorm(100, 0, 1)
#bout <- id_outliers(vec)
#dt <- data.table(Values = vec, bOutlier = bout)
#dt[bOutlier == TRUE]
}
