% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickstats.R
\name{MorphStat}
\alias{MorphStat}
\alias{plyStat}
\title{Summary Statistics}
\usage{
MorphStat(x, lab = "Value", d = 3, na.rm = TRUE)

plyStat(dt, cols)
}
\arguments{
\item{x}{An r expression naming the column to run a custom set of statistics functions on}

\item{lab}{A character value labeling a newly created column of results. Default is "Value"}

\item{d}{A numeric digit representing the number of digits to round in the final answer. Default is 3.}

\item{na.rm}{A boolean with a default of TRUE that indicates whether NAs should be removed from the 
calculation.}

\item{dt}{A data.table to apply MorphStat through}

\item{cols}{An optional argument for \code{plyStats} that subsets the data.table by column
prior to applying MorphStat functions}
}
\description{
A set of functions to run a set of customized summary statistics on 1 numeric column of a data.table or
all columns of a data.table that contains only numeric values.
}
\section{Functions}{
\itemize{
\item \code{MorphStat}: A function to apply a set of summary statistics function 
to a column in a data.table

\item \code{plyStat}: A wrapper that has vectorized \code{MorphStat} over a number of columns specified by the 
argument \code{x}
}}
\examples{
#
# Show examples here
#
}

