% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/easyfilter.R
\name{rFilter_All}
\alias{cFilter}
\alias{cRemoveNA}
\alias{rFilter}
\alias{rFilter_All}
\title{Quick filtering rows or columns}
\usage{
rFilter_All(dt, val = NA, boolfun = NULL, cols = NULL, copy = TRUE)

rFilter(dt, val = NA, exact = FALSE, set = NULL)

cRemoveNA(dt)

cFilter(dt, val = NA, boolfun = NULL, cols = NULL)
}
\arguments{
\item{dt}{A data.table to filter rows or columns}

\item{val}{A value to filter rows or columns on}

\item{boolfun}{A function that can be used instead of val to filter. This function
must be a "predicate" function applied across rows (\code{cFilter}), or down
columns (\code{rFilter_All}). See details and examples below.
(returns T/F based on a test of values in dt
applied across rows, or down columns)}

\item{cols}{An OPTIONAL character vector specifying columns to include as part of
testing the conditions. If cols is not provided, all column of dt will be
tested}

\item{copy}{A boolean indicating whether the returned data.table should be a
copy of "dt" or a reference to the data in memory. Default value is TRUE to protect
against unnexpected changes to the original data, however setting copy=FALSE when
appropriate is much more efficient. See example use cases.}

\item{exact}{A boolean that describes whether to exactly match. Default is FALSE which means a text based
search will occur in order to match whether val is contained in the row}

\item{set}{An optional list element that can be provided to set a desired new column after filtering.
set should be a named list where the name corresponds to the new column name, and the value
is the new column value. This is ideal when needing to quick boolean or indicator column to
identify the group that was just filtered. See examples}
}
\description{
\code{rFilter_All} is a function that enables quick filtering a data.table based on
 a TRUE/FALSE condition applied down the column (e.g. returns entire row if
 condition is TRUE for each value down rows of specified columns).
 \code{cFilter} is analogous for columns. This will test a condition across rows
 of every specified columns and drop each if the condition is met.
}
\section{Functions}{
\itemize{
\item \code{rFilter_All}: A function to filter rows of a data.table

\item \code{rFilter}: A function to apply a text based search to a given data table's rows and columns,
returning rows where any column's value meets the search condition

\item \code{cRemoveNA}: A function to filter columns of a data.table

\item \code{cFilter}: A function to filter columns of a data.table
}}

