% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIndex.R
\name{SplitIndex}
\alias{SplitIndex}
\title{Generate Indices}
\usage{
SplitIndex(totalN = NULL, eachN, mat = FALSE)
}
\arguments{
\item{eachN}{A numeric value representing the desired number of splits}

\item{mat}{A boolean indicating whether to return the splits as a list or matrix (simplified)}

\item{len}{A numeric value representing the total observations to split an index over}
}
\description{
A function to create a split index. Comes in handy if parallelizing a computation over a very large dataset
}
\examples{
len <- 1000
eachN <- 10
SplitIndex(len, eachN) # Not simplified
SplitIndex(len, eachN, mat = TRUE) # simplified
}

