% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xIndex.R
\name{SplitIndex}
\alias{SplitIndex}
\alias{splitn}
\title{Generate Indices}
\usage{
SplitIndex(totalN = NULL, eachN, mat = FALSE)

splitn(vec, n)
}
\arguments{
\item{totalN}{A numeric value representing the total observations to split an index over}

\item{eachN}{A numeric value representing the desired number of splits}

\item{mat}{A boolean indicating whether to return the splits as a list or matrix (simplified)}

\item{vec}{A vector to split into n list elements}

\item{n}{The number of desired list elements to split vec into}
}
\description{
A function to create a split index. Comes in handy if parallelizing a computation over a very large dataset
}
\section{Functions}{
\itemize{
\item \code{splitn}: A convenience wrapper around splitIndices. Reduces a step necessary if 
desire is to split a defined vector
}}
\examples{
totalN <- 1000
eachN <- 10
SplitIndex(totalN, eachN) # Not simplified
SplitIndex(totalN, eachN, mat = TRUE) # simplified
splitn(letters, 4)
}

